@@ -26,8 +26,7 @@ allprojects
 
   ext
   {
-    appcompat_support_library_version = '28.0.0'
-    compileSdkVersion = 28
+    compileSdkVersion = 29
     minSdkVersion = 14
     GROUP_ID = 'com.smartnsoft'
     ARTIFACT_ID = 'recyclerview'
