@@ -31,10 +31,10 @@ import android.app.Activity;
 import android.content.Context;
 import android.graphics.Point;
 import android.os.Build.VERSION_CODES;
-import android.support.annotation.UiThread;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.RecyclerView.Adapter;
+import androidx.annotation.UiThread;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView.Adapter;
 import android.util.SparseArray;
 import android.view.Display;
 import android.view.View;
