@@ -23,7 +23,7 @@
 package com.smartnsoft.recyclerview.attributes;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView.ViewHolder;
+import androidx.recyclerview.widget.RecyclerView.ViewHolder;
 import android.view.View;
 
 /**
