@@ -22,8 +22,8 @@
 
 package com.smartnsoft.recyclerview.attributes;
 
-import android.databinding.ViewDataBinding;
-import android.support.annotation.CallSuper;
+import androidx.databinding.ViewDataBinding;
+import androidx.annotation.CallSuper;
 
 /**
  * A {@link SmartRecyclerAttributes} that uses databinding.
