@@ -23,8 +23,8 @@
 package com.smartnsoft.recyclerview.recyclerview
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.helper.ItemTouchHelper
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.ItemTouchHelper
 import android.util.AttributeSet
 import com.smartnsoft.recyclerview.adapter.SmartRecyclerAdapter
 
