@@ -1,7 +1,7 @@
 package com.smartnsoft.smartrecyclerviewsample.wrappers;
 
 import android.content.Context;
-import android.support.annotation.DrawableRes;
+import androidx.annotation.DrawableRes;
 import android.view.View;
 
 import com.smartnsoft.smartrecyclerviewsample.R;
